requirements.txt
discord.py
requests
psutil
neuralintents